Update Breadth-First Search (BFS).cpp

+++ Algorithms-and-Common-Problems/Graph/Breadth-First Search (BFS).cpp
@@ -2,13 +2,13 @@ const ll N = 1e5+7;
 vector <ll> adjList[N];
 bool visited[N];
 
-void bfs (int rootNode) {
-    queue <int> q;
+void bfs (ll rootNode) {
+    queue <ll> q;
     q.push(rootNode);
     visited[rootNode] = true;
 
     while (!q.empty()) {
-        int currNode = q.front();
+        ll currNode = q.front();
         q.pop();
 
         for (auto &u : adjList[currNode]) {
